ref: extract business logic into useDatePicker hook
